test(relay): add apply_config tests and consolidate module tests
